[Project]: moved important core files to 'core/' dir.
