+++ src/ts/classes/core/Primitive.ts
@@ -1,4 +1,4 @@
-import { changeColorOpacity } from "../helpers";
+import { changeColorOpacity } from "../../helpers";
 import RectangleShape from "./Rectangle";
 import SynteticEventTarget from "./SynteticEventTarget";
