+++ src/ts/classes/core/Rectangle.ts
