+++ src/ts/classes/core/SynteticEventTarget.ts
